Spike/docs testing (#325)
# Description

Adds the two scripts used for testing.

Also increase a timeout because the test failed in azure. Example loading is really slow, I think it's because the load job has no requests. I made a ticket for it https://github.com/stackabletech/superset-operator/issues/324

While I looked at it, the CPU usage of the python load process was really high, so that's why I think it might be that.

https://ci.stackable.tech/view/05%20Documentation%20Tests/job/superset-operator-documentation-test/3/