chore: Add support Superset 6.0.0-rc2 (#680)
* fix: Remove deprecated and removed import

Flask AppBuilder 5 no longer has the AUTH_OID module. It is possible a user still relies on this for old Superset versions.

* fix: Set required config (to an empty value

Flask AppBuilder 5, for some reason, requires `RECAPTCHA_PUBLIC_KEY` to be set, or the UI cannot be used, and the following error is emitted on the container console:

```
  File "/stackable/app/lib64/python3.11/site-packages/superset/views/base.py", line 417, in cached_common_bootstrap_data
    frontend_config["RECAPTCHA_PUBLIC_KEY"] = app.config["RECAPTCHA_PUBLIC_KEY"]
                                              ~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^^^
KeyError: 'RECAPTCHA_PUBLIC_KEY'
```

* chore: Use to_owned for &str -> String

* test(kuttl): Add Superset 6.0.0-rc2 to test-definitions.

This requires the product image to be merged and built: https://github.com/stackabletech/docker-images/pull/1337

* chore: Update changelog

* docs: Update getting started

* docs: Update supported versions

By SDP 26.3, this should no longer be rc2

* chore: Update changelog

* docs: Update required-external-components